Added Excaption handling

+++ src/main/java/com/maveric/transactionservice/service/TransactionServiceImp.java
@@ -1,37 +1,47 @@
 package com.maveric.transactionservice.service;
+
 import com.maveric.transactionservice.dto.TransactionDto;
 import com.maveric.transactionservice.exception.TranscationNotFoundException;
 import com.maveric.transactionservice.model.Transaction;
 import com.maveric.transactionservice.repository.TransactionRepository;
 import com.maveric.transactionservice.userDefinedMethodsAndConstants.Methods;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpServerErrorException;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import static com.maveric.transactionservice.userDefinedMethodsAndConstants.Methods.dto;
-import static com.maveric.transactionservice.userDefinedMethodsAndConstants.Methods.getCurrentDateTime;
 
 @Service
 public class TransactionServiceImp implements TransactionService{
     @Autowired
-    private TransactionRepository repository;
+     TransactionRepository repository;
+    String exception ="Transaction Not Found";
     @Override
-    public List<TransactionDto> getTransactions(Integer page,Integer pageSize) {
-        Page<Transaction> transactions=repository.findAll(PageRequest.of(page,pageSize));
-        List<TransactionDto> transactionDto= new ArrayList<>();
-        for (Transaction a:transactions){
-            transactionDto.add(dto(a));
-        }
-        return transactionDto;
+    public List<TransactionDto> getTransactions(@NotNull String accountId, Integer page, Integer pageSize) throws TranscationNotFoundException {
+        Page<Transaction> transactions = repository.findAll(PageRequest.of(page, pageSize));
+        List<TransactionDto> transactionDto = new ArrayList<>();
+        if (accountId.isEmpty()) {
+            throw new TranscationNotFoundException(exception);
+        } else {
+            for (Transaction a : transactions) {
+                if (a.getAccountId().contentEquals(accountId)) {
+                    transactionDto.add(dto(a));
+                   }
 
-    }
+            }
+        } return transactionDto;}
+
+
     @Override
-    public TransactionDto createTransaction(TransactionDto transactionDto) {
+    public TransactionDto createTransaction(TransactionDto transactionDto) throws HttpServerErrorException.InternalServerError {
        Transaction transaction = new Transaction();
+
        transaction.set_id(transactionDto.get_id());
        transaction.setAccountId(transactionDto.getAccountId());
        transaction.setAmount(transactionDto.getAmount());
@@ -42,17 +52,18 @@ public class TransactionServiceImp implements TransactionService{
     }
 
     @Override
-    public TransactionDto getTransactionById(String transactionDtoId) {
-        Transaction transactionResponse=repository.findById(transactionDtoId).orElseThrow(() -> new TranscationNotFoundException("Transaction not found"));
-        return dto(transactionResponse);
-    }
+    public TransactionDto getTransactionById(String accountDtoId, String transactionDtoId) throws TranscationNotFoundException{
+        Transaction response = repository.findById(transactionDtoId).orElseThrow(() -> new TranscationNotFoundException(exception));
+        if(response.getAccountId().contentEquals(accountDtoId)){
+       return dto(response);}else{throw new TranscationNotFoundException(exception);}
+       }
 
     @Override
-    public String deleteTransaction(String transactionDtoId) {
-        repository.deleteById(transactionDtoId);
-        return "Transaction deleted successfully.";
+    public void deleteTransaction(String accountId,String transactionDtoId) throws TranscationNotFoundException {
+        Transaction transaction = repository.findById(transactionDtoId).orElseThrow(() ->new TranscationNotFoundException("Transaction Not Found"));
+        if (transaction.getAccountId().contains(accountId)){
+            repository.deleteById(transactionDtoId);
+        }else { throw new TranscationNotFoundException(exception);}
     }
-
-
 
 }
